fixed namedQueries, UserInfoTO fix, UserShowService tests

+++ src/main/java/com/capgemini/chess/dataaccess/dao/UserDao.java
@@ -10,6 +10,6 @@ public interface UserDao extends Dao<UserTO, Long> {
 	
 	List<UserTO> findAllAndOrderByPoints();
 	
-	int countUsersWithPointsMoreThan(int points);
+	long countUsersWithPointsMoreThan(int points);
 
 }
